+++ cfg/validator/non_finish_places_test.go
@@ -1,3 +1,4 @@
+// nolint:funlen
 package validator
 
 import (